requirements.txt
google-generativeai
Flask
algosdk
msgpack